# BitWaves.WebAPI/docs/swagger/bitwaves-v1.yaml
swagger: '2.0'
info:
  description: BitWaves 后端接口文档
  version: 1.0.0
  title: BitWaves Web API Docs
  contact:
    email: msrlancern@126.com
...
paths:
...
  /problems:
    get:
      summary: 获取题目列表
      description: 获取题目列表
      tags:
        - problems
      parameters:
        - in: header
          name: Authorization
          type: string
          required: true
          description: 用户身份验证标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: query
          name: page
          type: integer
          required: false
          description: 分页参数，页面编号。页面编号从 0 开始
        - in: query
          name: itemsPerPage
          type: integer
          required: false
          description: 分页参数，每一页上的元素数量。默认为 20
      responses:
        200:
          description: 操作成功完成
          headers:
            X-BitWaves-Count:
              type: integer
              description: 分页前满足筛选条件的题目数量
          schema:
            type: array
            items:
              $ref: '#/definitions/ProblemListItem'
        400:
          description: 请求中携带的参数无效
        403:
          description: 用户没有相应的权限
    post:
      summary: 创建一个题目
      description: 创建一个题目
      tags:
        - problems
      parameters:
        - in: header
          name: Authorization
          type: string
          required: true
          description: 身份验证标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: body
          name: data
          required: true
          description: 题目创建数据
          schema:
            $ref: '#/definitions/PostProblemModel'
      responses:
        200:
          description: 操作成功完成
          schema:
            type: object
            required:
              - id
            properties:
              id:
                type: string
                description: 新创建的题目的全局唯一 ID
                example: "a154edb2431fdbb230fbf384"
        400:
          description: 给定的题目创建数据无效
        403:
          description: 用户没有权限创建题目
  /problems/{problemId}:
    put:
      summary: 修改题目信息
      description: 修改题目信息
      tags:
        - problems
      parameters:
        - in: path
          name: problemId
          type: string
          required: true
          description: 题目的全局唯一 ID
        - in: header
          name: Authorization
          type: string
          required: true
          description: 身份验证头部字段，内容格式为 `Jwt <jwt>`，注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: body
          name: data
          required: true
          description: 题目修改信息
          schema:
            $ref: '#/definitions/PutProblemModel'
      responses:
        200:
          description: 操作成功完成
        400:
          description: 给定的题目修改信息无效
        403:
          description: 用户没有权限修改题目信息
        404:
          description: 题目不存在
    get:
      summary: 获取题目信息
      description: 获取题目信息
      tags:
        - problems
      parameters:
        - in: path
          name: problemId
          required: true
          type: string
          description: 题目的全局唯一 ID
        - in: header
          name: Authorization
          required: false
          type: string
          description: 当访问的题目不是公开题目时，提供用户身份验证标识 JWT。该字段的内容格式为 `Jwt <jwt>`，注意 `Jwt` 与 `<jwt>` 之间的空格
      responses:
        200:
          description: 操作成功完成
          schema:
            $ref: '#/definitions/ProblemInfo'
        400:
          description: 无效的题目 ID
        403:
          description: 用户没有足够的权限访问题目
        404:
          description: 未能找到指定的题目
  /problems/{problemId}/tags:
    post:
      summary: 向题目添加标签
      description: 向题目添加标签
      tags:
        - problems
      parameters:
        - in: header
          name: Authorization
          type: string
          required: true
          description: 用户身份标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: body
          name: data
          required: true
          description: 要添加的标签名称
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: 操作成功完成
        400:
          description: 请求中包含无效的标签名称
        403:
          description: 用户没有相应的权限
        404:
          description: 无法找到要求的题目
    delete:
      summary: 从题目删除标签
      description: 从题目删除标签
      tags:
        - problems
      parameters:
        - in: header
          name: Authorization
          type: string
          required: true
          description: 用户身份标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: body
          name: data
          required: true
          description: 要删除的标签名称
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: 操作成功完成
        400:
          description: 请求中包含无效的标签名称
        403:
          description: 用户没有相应的权限
        404:
          description: 无法找到要求的题目
  /problems/tags:
    get:
      summary: 获取所有的题目标签
      description: 获取所有的题目标签
      tags:
        - problemTags
      parameters:
        - in: header
          name: Authorization
          required: true
          type: string
          description: 用户身份标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
      responses:
        200:
          description: 操作成功完成
          schema:
            type: array
            items:
              type: string
            example:
              - 暴力
              - 贪心
              - 数据结构
        403:
          description: 用户没有相应的权限
...
  /languages:
    get:
      summary: 获取所有可用的评测语言列表
      description: 获取所有可用的评测语言列表
      tags:
        - languages
      responses:
        200:
          description: 操作成功完成
          schema:
            type: array
            items:
              $ref: '#/definitions/LanguageInfo'
    post:
      summary: 创建一个新评测语言
      description: 创建一个新评测语言
      tags:
        - languages
      parameters:
        - in: header
          name: Authorization
          type: string
          required: true
          description: 用户身份验证标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: body
          name: data
          required: true
          description: 要创建的语言数据
          schema:
            $ref: '#/definitions/PostLanguageModel'
      responses:
        200:
          description: 操作成功完成
          schema:
            type: object
            required:
              - id
            properties:
              id:
                type: string
                description: 新创建的语言的全局唯一 ID
                example: "a154edb2431fdbb230fbf384"
        400:
          description: 无效的请求数据
        403:
          description: 用户不具备需要的权限
  /languages/{languageId}:
    delete:
      summary: 删除评测语言
      description: 删除评测语言
      tags:
        - languages
      parameters:
        - in: header
          name: Authorization
          type: string
          required: true
          description: 用户身份验证标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: path
          name: languageId
          type: string
          required: true
          description: 要删除的评测语言的全局唯一 ID
      responses:
        200:
          description: 操作成功完成
        400:
          description: 无效的评测语言 ID
        403:
          description: 用户没有相应的权限
        404:
          description: 未能找到指定的评测语言
definitions:
...
  PostProblemModel:
    type: object
    required:
      - title
    properties:
      title:
        type: string
        minLength: 1
        maxLength: 128
        description: 题目标题
        example: "诸神眷顾的幻想乡"
  PutProblemModel:
    type: object
    properties:
      title:
        type: string
        minLength: 1
        maxLength: 128
        description: 题目的标题。若不需要修改，请不要提供此字段
        example: "诸神眷顾的幻想乡"
      legend:
        type: string
        description: 题目的背景描述。若不需要修改，请不要提供此字段
        example: "hello world"
      input:
        type: string
        description: 题目的输入描述。若不需要修改，请不要提供此字段
        example: "input description"
      output:
        type: string
        description: 题目的输出描述。若不需要修改，请不要提供此字段
        example: "output description"
      notes:
        type: string
        description: 题目的提示信息。若不需要修改，请不要提供此字段
        example: "notes"
      difficulty:
        type: integer
        description: 题目的难度系数。若不需要修改，请不要提供此字段
        example: 80
      timeLimit:
        type: integer
        minimum: 500
        maximum: 10000
        description: 单个测试点的时间限制，单位为毫秒。若不需要修改，请不要提供此字段
        example: 1000
      memoryLimit:
        type: integer
        minimum: 32
        maximum: 1024
        description: 单个测试点的内存限制，单位为MB。若不需要修改，请不要提供此字段
        example: 256
      judgeMode:
        type: string
        description: 题目的评测模式。若不需要修改，请不要提供此字段
        enum:
          - Standard
          - SpecialJudge
          - Interactive
        example: SpecialJudge
      builtinCheckerOptions:
        type: array
        items:
          type: string
        description: 当评测模式为 Standard 模式时，该成员指出传递给内建答案检查器的参数。可能的值包括：`FloatingPointAware` 以及 `IgnoreCase`
        example:
          - FloatingPointAware
  ProblemListItem:
    type: object
    required:
      - id
      - archiveId
      - title
      - creationTime
      - lastUpdateTime
      - author
      - source
      - difficulty
      - tags
      - totalSubmissions
      - acceptedSubmissions
      - totalAttemptedUsers
      - totalSolvedUsers
      - lastSubmissionTime
    properties:
      id:
        type: string
        description: 题目的全局唯一 ID，请注意与题目的 archiveId 的区别
        example: "5d8825e87d2847a7e4886fbf"
      archiveId:
        type: integer
        description: 题目在 Problem Archive 中的 ID
        example: 10
      title:
        type: string
        description: 题目的标题
        example: "Two Sums"
      creationTime:
        type: string
        description: 题目的创建时间，UTC 时区
        example: "2019-09-23T01:54:48.087Z"
      lastUpdateTime:
        type: string
        description: 题目上次更新的时间，UTC 时区
        example: "2019-09-23T01:54:48.087Z"
      author:
        type: string
        description: 题目的作者的用户名
        example: "admin"
      source:
        type: string
        description: 题目的来源
        example: "2018年北京理工大学程序设计校赛"
      difficulty:
        type: integer
        description: 题目的难度系数
        example: 80
      tags:
        type: array
        items:
          type: string
        description: 题目的标签
        example:
          - 字符串
          - 模拟
          - 数据结构
      totalSubmissions:
        type: integer
        description: 题目的总提交数量
        example: 1000
      acceptedSubmissions:
        type: integer
        description: 题目的 AC 提交数量
        example: 100
      totalAttemptedUsers:
        type: integer
        description: 尝试过该题目的用户总数
        example: 100
      totalSolvedUsers:
        type: integer
        description: 成功解答该题目的用户总数
        example: 40
      lastSubmissionTime:
        type: string
        description: 题目的上次提交时间
        example: "2019-09-23T01:54:48.087Z"
  ProblemInfo:
    type: object
    required:
      - id
      - archiveId
      - title
      - creationTime
      - lastUpdateTime
      - author
      - source
      - legend
      - input
      - output
      - samples
      - notes
      - difficulty
      - tags
      - timeLimit
      - memoryLimit
      - judgeMode
      - totalSubmissions
      - acceptedSubmissions
      - testReady
    properties:
      id:
        type: string
        description: 题目的全局唯一 ID
        example: "5d8825e87d2847a7e4886fbf"
      archiveId:
        type: integer
        description: 题目在公开题目集中的 ID。若题目不在公开题目集中，该字段为 null
      title:
        type: string
        description: 题目标题
        example: "诸神眷顾的幻想乡"
      creationTime:
        type: string
        description: 题目的创建时间，UTC 时区
        example: "2019-09-23T01:54:48.087Z"
      lastUpdateTime:
        type: string
        description: 题目上次更新的时间，UTC 时区
        example: "2019-09-23T01:54:48.087Z"
      author:
        type: string
        description: 题目的作者的用户名
        example: "admin"
      source:
        type: string
        description: 题目的来源
        example: "2018年北京理工大学程序设计校赛"
      legend:
        type: string
        description: 题目正文叙述
        example: "hello world"
      input:
        type: string
        description: 题目输入描述
        example: "input description"
      output:
        type: string
        description: 题目输出描述
        example: "output description"
      samples:
        type: array
        description: 题目样例
        items:
          type: object
          properties:
            input:
              type: string
              description: 样例输入
              example: "1 2"
            output:
              type: string
              description: 样例输出
              example: "3"
      notes:
        type: string
        description: 题目提示信息
        example: "notes"
      difficulty:
        type: integer
        description: 题目难度系数，范围为 [0, 100]
        example: 80
      tags:
        type: array
        description: 题目标签
        items:
          type: string
        example:
          - "字符串"
          - "模拟"
          - "数据结构"
      timeLimit:
        type: integer
        description: 题目时间限制，单位为毫秒
        example: 1000
      memoryLimit:
        type: integer
        description: 题目内存限制，单位为 MB
        example: 256
      judgeMode:
        type: string
        description: 题目评测模式
        enum:
          - Standard
          - SpecialJudge
          - Interactive
      builtinCheckerOptions:
        type: array
        items:
          type: string
        description: 当评测模式为 Standard 时，该字段表示内建答案检查器的选项。可能的值有 `FloatingPointAware` 以及 `IgnoreCase`
        example:
          - FloatingPointAware
          - IgnoreCase
      totalSubmissions:
        type: integer
        description: 题目总提交数量
        example: 100
      acceptedSubmissions:
        type: integer
        description: 题目 AC 提交数量
        example: 80
      totalAttemptedUsers:
        type: integer
        description: 尝试过该题目的用户总数
        example: 100
      totalSolvedUsers:
        type: integer
        description: 成功解答该题目的用户总数
        example: 40
      lastSubmissionTime:
        type: string
        description: 题目的上次提交时间
        example: "2019-09-23T01:54:48.087Z"
      testReady:
        type: boolean
        description: 用户是否可以提交答案
        example: true
...
  PostLanguageModel:
    type: object
    required:
      - langId
      - dialect
      - version
      - displayName
    properties:
      langId:
        type: string
        minLength: 1
        maxLength: 32
        description: 语言标识符
        example: cpp
      dialect:
        type: string
        minLength: 1
        maxLength: 32
        description: 方言标识符
        example: clang
      version:
        type: string
        minLength: 1
        maxLength: 32
        description: 版本标识符
        example: '17'
      displayName:
        type: string
        minLength: 1
        maxLength: 128
        description: 语言的显示名称
        example: 'Clang++ 17'
  LanguageInfo:
    type: object
    required:
      - id
      - langId
      - dialect
      - version
      - displayName
    properties:
      id:
        type: string
        description: 语言的全局唯一标识符
        example: '25fa137e2b3513ace3253cdf'
      langId:
        type: string
        description: 语言标识符
        example: cpp
      dialect:
        type: string
        description: 方言标识符
        example: clang
      version:
        type: string
        description: 版本标识符
        example: '17'
      displayName:
        type: string
        description: 语言的显示名称
        example: 'Clang++ 17'